+++ arcade/gui/examples/style_v2.py
@@ -8,6 +8,7 @@ is disabled by the user.
 If arcade and Python are properly installed, you can run this example with:
 python -m arcade.gui.examples.style_v2
 """
+
 from __future__ import annotations
 
 from itertools import cycle
@@ -132,7 +133,7 @@ class MyView(arcade.View):
         self.ui.draw()
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     window = arcade.Window(800, 600, "UIExample", resizable=True)
     window.show_view(MyView())
     window.run()