+++ tests/unit/gui/test_uilabel.py
@@ -1,14 +1,15 @@
+from unittest.mock import Mock
+
 import pytest
 
 from arcade.gui import UILabel, Rect
+from arcade.types import Color
 
 
 def test_uilabel_inits_with_text_size(window):
     label = UILabel(text="Example")
 
-    assert label.rect.width == pytest.approx(
-        63, abs=6
-    )  # on windows the width differs about 6 pixel
+    assert label.rect.width == pytest.approx(63, abs=6)  # on windows the width differs about 6 pixel
     assert label.rect.height == pytest.approx(19, abs=1)
 
 
@@ -56,6 +57,7 @@ def test_uilabel_fixes_internal_text_to_pos_0_0(window):
     assert label.label.position == (0, 0)
     assert label.position == (10, 10)
 
+
 def test_adaptive_width_support_for_multiline_text(window):
     """
     This test is a bit tricky. Enabling multiline without a width
@@ -65,3 +67,31 @@ def test_adaptive_width_support_for_multiline_text(window):
     label = UILabel(text="Multiline\ntext\nwhich\n", multiline=True)
     assert label.width < 100
     assert label.height > 20
+
+
+def test_change_text_does_full_render_without_background(window):
+    """
+    This test is a bit tricky. Enabling multiline without a width
+    should fit the size to the text. This is not natively supported by either arcade.Text or pyglet.Label.
+    Because text length variates between different os, we can only test boundaries, which indicate a proper implementation.
+    """
+
+    label = UILabel(text="First Text")
+    label.parent = Mock()
+
+    label.text = "Second Text"
+    label.parent.trigger_render.assert_called_once()
+
+
+def test_change_text_does_normal_render_with_background(window):
+    """
+    This test is a bit tricky. Enabling multiline without a width
+    should fit the size to the text. This is not natively supported by either arcade.Text or pyglet.Label.
+    Because text length variates between different os, we can only test boundaries, which indicate a proper implementation.
+    """
+
+    label = UILabel(text="First Text").with_background(color=Color(255, 255, 255, 255))
+    label.parent = Mock()
+
+    label.text = "Second Text"
+    label.parent.trigger_render.assert_not_called()